+++ postgres_sql_queries/query_067.sql
@@ -7,7 +7,7 @@
 -- Extracted on: 2025-05-25T17:11:39.335585
 
 SELECT		locationext_id, 
-				redhires, 
+				redhires::integer, 
 				TO_CHAR(grandopeningdate, 'YYYY-MM-DD HH24:MI:SS') as grandopeningdate,  
 				<BATCHDATE> as ExtractDate, 
 				<BATCHDATE> as BatchDate,
